requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
multidict==6.0.2
psycopg2-binary==2.9.3
pydantic==1.9.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
yarl==1.7.2
zipp==3.8.0
